Fixing bug on Address and Credit Card while editing
Bug Fixed: when the user updates a credit card or an address, creates a new row in database with the new data.

Problem: the object is passed without an ID, so Hibernate understands it as new data to save.

Solution: created hidden inputs with the address and credit card in each form to Hibernate have a data reference to merge.